refactor: move HttpHeadersConstants to common-constants
