+++ gateway/src/main/java/ru/practicum/shareit/itemrequest/controller/ItemRequestController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.*;
 import ru.practicum.shareit.commondto.itemrequest.request.ItemRequestCreationDto;
 import ru.practicum.shareit.commondto.itemrequest.response.ItemRequestDto;
 import ru.practicum.shareit.commondto.validation.Groups;
-import ru.practicum.shareit.constants.HttpHeadersConstants;
+import ru.practicum.shareit.commons.constants.HttpHeadersConstants;
 import ru.practicum.shareit.itemrequest.client.ItemRequestClient;
 
 import javax.validation.constraints.Positive;
